requirements.txt
Flask
Pillow
pytesseract
pymongo
python-dotenv
supabase
